Trading/crypto_bot/requirements.txt
# ──────────────────────────────────────────────────────────────────
# Crypto Auto Trading Bot — Python Dependencies
# ──────────────────────────────────────────────────────────────────

# Exchange connectivity
ccxt>=4.0.0,<5.0.0

# Data / numerics
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<3.0.0

# Environment / config
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
